refactor: export common graphs functions.
