+++ frontend/src/app/components/mempool-graph/mempool-graph.component.ts
@@ -7,6 +7,7 @@ import { StateService } from 'src/app/services/state.service';
 import { StorageService } from 'src/app/services/storage.service';
 import { EChartsOption } from '@mempool/echarts';
 import { feeLevels, chartColors } from 'src/app/app.constants';
+import { dataZoom, formatterXAxis, labelsData, filterLabelDates } from 'src/app/shared/graphs.utils';
 
 @Component({
   selector: 'app-mempool-graph',
@@ -99,95 +100,14 @@ export class MempoolGraphComponent implements OnInit, OnChanges {
     return finalArray;
   }
 
-  addMinutes =  function (dt: Date, minutes: number) {
-    return new Date(dt.getTime() + minutes * 60000);
-  }
-
-  filterLabelDates(labels: any) {
-    let labelDates = [];
-
-    if (labels.length > 0) {
-      const firstDate = new Date(labels[0]);
-      const lastDate = new Date(labels[labels.length - 1]);
-      let currentDate = firstDate;
-
-      currentDate.setUTCMilliseconds(0);
-      currentDate.setUTCSeconds(0);
-      currentDate.setUTCMinutes(0);
-
-      if (this.template !== 'widget') {
-
-        if(['1w', '1m', '2m', '3m', '6m', '1y', '2y', '3y'].includes(this.windowPreference)){
-          currentDate.setUTCMinutes(0);
-        }
-
-        if(['6m', '1y', '2y', '3y'].includes(this.windowPreference)){
-          currentDate.setUTCHours(0);
-        }
-      }
-
-      let timeIntervals = {
-        '2h' : 10,           // 10 mins
-        '24h': 120,          //  2 hours
-        '1w' : 60 * 24,      //  1 day
-        '1m' : 60 * 24 * 7,  //  7 days
-        '3m' : 60 * 24 * 7,  //  7 days
-        '6m' : 60 * 24 * 15, // 15 days
-        '1y' : 60 * 24 * 30, // 30 days
-        '2y' : 60 * 24 * 60, // 30 days
-        '3y' : 60 * 24 * 60, // 30 days
-      };
-
-      if (window.innerWidth < 600 || this.template === 'widget') {
-        timeIntervals = {
-          '2h' : 30,           // 30 mins
-          '24h': 60 * 6,       //  6 hours
-          '1w' : 60 * 48,      //  2 day
-          '1m' : 60 * 24 * 7,  //  7 days
-          '3m' : 60 * 24 * 7,  // 15 days
-          '6m' : 60 * 24 * 30, // 30 days
-          '1y' : 60 * 24 * 30, // 30 days
-          '2y' : 60 * 24 * 60, // 30 days
-          '3y' : 60 * 24 * 60, // 30 days
-        };
-      }
-
-      const windowPreference = this.windowPreference in timeIntervals && this.template !== 'widget' ? this.windowPreference : '2h';
-      while (currentDate.getTime() < lastDate.getTime()) {
-        currentDate = this.addMinutes(currentDate, timeIntervals[windowPreference]);
-        labelDates.push(currentDate.toISOString());
-      }
-    }
-
-    return labelDates;
-  }
-
   mountFeeChart() {
     this.orderLevels();
     const { labels, series } = this.mempoolVsizeFeesData;
-    const labelsData = labels.map((label: string, index: number) => {
-      const date = new Date(label);
-
-      date.setUTCMilliseconds(0);
-      date.setUTCSeconds(0);
-
-      if (this.template !== 'widget') {
-
-        if(['1w', '1m', '2m', '3m', '6m', '1y', '2y', '3y'].includes(this.windowPreference)){
-          date.setUTCMinutes(0);
-        }
-
-        if(['6m', '1y', '2y', '3y'].includes(this.windowPreference)){
-          date.setUTCHours(0);
-        }
-      }
-
-      return date.toISOString();
-    })
+    const labelsDataLocal = labelsData(labels, this.template, this.windowPreference);
 
     const seriesGraph = [];
     const newColors = [];
-    const labelsFiltered = this.filterLabelDates(labelsData);
+    const labelsFiltered = filterLabelDates(labelsDataLocal, this.template, this.windowPreference);
 
     for (let index = 0; index < series.length; index++) {
       const value = series[index];
@@ -356,44 +276,7 @@ export class MempoolGraphComponent implements OnInit, OnChanges {
           </div>`;
         }
       },
-      dataZoom: [{
-        type: 'inside',
-        realtime: true,
-        zoomOnMouseWheel: (this.template === 'advanced') ? true : false,
-        maxSpan: 100,
-        minSpan: 10,
-      }, {
-        show: (this.template === 'advanced' && this.showZoom) ? true : false,
-        type: 'slider',
-        brushSelect: false,
-        realtime: true,
-        bottom: 0,
-        labelFormatter: (value, valueStr) => {
-          const date = new Date (valueStr);
-          switch (this.windowPreference) {
-            case '1w':
-            case '1m':
-              return date.toLocaleDateString(this.locale, { month: 'short', weekday: 'short', day: 'numeric' });
-            case '3m':
-            case '6m':
-            case '1y':
-            case '2y':
-            case '3y':
-              return date.toLocaleDateString(this.locale, { year: 'numeric', month: 'short' });
-            default: // 2m, 24h
-              return date.toLocaleTimeString(this.locale, { hour: 'numeric', minute: 'numeric' });
-          }
-        },
-        selectedDataBackground: {
-          lineStyle: {
-            color: '#fff',
-            opacity: 0.45,
-          },
-          areaStyle: {
-            opacity: 0,
-          }
-        }
-      }],
+      dataZoom: dataZoom(this.template, this.locale, this.showZoom, this.windowPreference),
       animation: false,
       grid: {
         height: this.height,
@@ -406,7 +289,7 @@ export class MempoolGraphComponent implements OnInit, OnChanges {
           type: 'category',
           boundaryGap: false,
           axisLine: { onZero: true },
-          data: labelsData,
+          data: labelsDataLocal,
           axisTick: {
             lineStyle: {
               width: 1,
@@ -419,33 +302,8 @@ export class MempoolGraphComponent implements OnInit, OnChanges {
             fontSize: 11,
             lineHeight: 25,
             customValues: ['1y', '2y', '3y'].includes(this.windowPreference) ? null : labelsFiltered,
-            formatter: (value: string) => {
-
-              if(value.length === 0){
-                return null;
-              }
-
-              const date = new Date(value);
-              if (this.template === 'widget') {
-                return date.toLocaleTimeString(this.locale, { hour: 'numeric', minute: 'numeric' });
-              }
-
-              switch (this.windowPreference) {
-                case '2h':
-                  return date.toLocaleTimeString(this.locale, { hour: 'numeric', minute: 'numeric' });
-                case '24h':
-                  return date.toLocaleTimeString(this.locale, { hour: 'numeric' });
-                case '1w':
-                case '1m':
-                  return date.toLocaleDateString(this.locale, { month: 'short', day: 'numeric', weekday: 'short' });
-                case '3m':
-                case '6m':
-                  return date.toLocaleDateString(this.locale, { month: 'short', day: 'numeric' });
-                case '1y':
-                case '2y':
-                case '3y':
-                  return date.toLocaleDateString(this.locale, { year: 'numeric', month: 'short' });
-              }
+            formatter: (value: any) => {
+              return formatterXAxis(this.template, this.locale, this.windowPreference, value );
             }
           },
         }
